Pathfinding: working in setPathingUnit() better
AlienBAIState: setPathingUnit() in each function before making calls to
	Pathfinding
CivilianBAIState: ditto.
BattlescapeGame::handleAI()
	- setPathingUnit() although it might be redundant w/ further
	Pathfinding calls from both/either AlienBAIState and UnitWalkBState
	- handlePanickingUnit() working on berserk routine ... mid-progress
	- moveUpDown() allows ALT to fall when flying
Map::cacheUnit() - fix to work w/ new Pathing algo.
Pathfinding: cast colors as Uint8 not int
	- remove setPathingUnit() call from calculate()
		- always setPathingUnit() independently.
	- might still need effort when passing in a missileTarget ....
	- previewPath() refactor.
	- validateUpDown() allows ALT to fall when flying
Projectile: fix no bullet sprite warning
UnitSprite: remove _moveType, was conflicting w/ setPathingUnit()
UnitWalkBState: call setPathingUnit() in cTor.
Logger: conform formatting etc.
BattleUnit: desc's + const
SavedBattleGame::setUnitPosition()
	- make sure setPathingUnit() gets a call before Pathinding funcs are
Tile: function name change, param cast change, function order change, desc.

+++ src/Battlescape/Pathfinding.h
@@ -111,7 +111,7 @@ private:
 			DIR_DOWN	=  9,
 			O_BIGWALL	= -1;
 
-		static int
+		static Uint8
 			red,
 			green,
 			yellow;
@@ -147,7 +147,6 @@ private:
 		/// Determines whether or not movement between starttile and endtile is possible in the direction.
 		bool isBlockedPath(
 				const Tile* const startTile,
-//				const Tile* endTile,
 				const int dir,
 				const BattleUnit* const missileTarget = NULL) const;
 
@@ -173,7 +172,6 @@ private:
 				const Position& startPos,
 				const int dir,
 				Position* destPos,
-//				BattleUnit* const unit,
 				const BattleUnit* const missileTarget = NULL,
 				bool missile = false);
 		/// Gets _totalTuCost; finds out whether we can hike somewhere in this turn or not.
@@ -182,7 +180,6 @@ private:
 
 		/// Checks if the movement is valid, for the up/down button.
 		int validateUpDown(
-//				const BattleUnit* const unit,
 				Position startPos,
 				int const dir);
